Fix crash on exit if Open3DScene is assigned to a python variable
